requirements.txt
python-dotenv
langchain==0.0.148
llama-index==0.5.6
ipython
requests
flask